fix: First refactory. Figured out how to combine pinata and pinata-browser
