+++ src/lib/providers/pinata-provider.ts
@@ -1,10 +1,11 @@
 import { PinataConfig, PinataPinOptions } from '@pinata/sdk';
+import FormData from 'form-data';
 
 import {
   uploadToIPFS,
   validateMetadata,
   validatePinataOptions,
-} from '../helpers/pinata-helpers-browser';
+} from '../helpers/pinata-provider-helpers';
 import { AssetBuffer } from '../interfaces/metadata';
 import { UploadProvider } from '../interfaces/profile-upload-options';
 
@@ -15,13 +16,13 @@ import { UploadProvider } from '../interfaces/profile-upload-options';
  * @param pinOptions pin options to use for all uploads
  * @returns <URL> return URL of uploaded file
  */
-export function createPinataBrowserUploader(
+export function createPinataUploader(
   pinataConfig: PinataConfig,
   pinOptions?: PinataPinOptions
 ): UploadProvider<PinataPinOptions> {
   const fn = async (data: any, options?: PinataPinOptions): Promise<URL> => {
     let meta: { 'content-type': string } | undefined;
-    const dataContent = new FormData();
+    const dataContent = new (global.FormData || FormData)();
     if (!('on' in data) && typeof data !== 'string') {
       if ('size' in data && 'type' in data) {
         const blob = data;
@@ -30,9 +31,9 @@ export function createPinataBrowserUploader(
       } else if ('buffer' in data && 'mimeType' in data) {
         const assetBuffer = data as AssetBuffer;
         meta = { 'content-type': assetBuffer.mimeType };
-        dataContent.append('file', new Blob([assetBuffer.buffer]));
+        dataContent.append('file', new (global.Blob || Blob)([assetBuffer.buffer]));
       } else if (Buffer.isBuffer(data)) {
-        dataContent.append('file', new Blob([data]));
+        dataContent.append('file', new (global.Blob || Blob)([data]));
       } else if ('on' in data && 'pipe' in data) {
         dataContent.append('file', data);
       } else {
